REF: splitting init file
