+++ lua/crookedtheme/commands.lua
@@ -1,8 +1,6 @@
-print("hello world");
+local crooked = require("crookedtheme.functions")
 
---local crooked = require("crookedtheme.init")
-
---[[ vim.api.nvim_create_user_command("Theme", 
+vim.api.nvim_create_user_command("Theme", 
     function(opts)
         crooked.Apply(opts.args)
     end,
@@ -10,10 +8,10 @@ print("hello world");
 )
 
 vim.api.nvim_create_user_command("ListThemes", 
-    function (opts)
-        for k,v in pairs(crooked.GetThemes()) do
+    function ()
+        for k,_ in pairs(crooked.GetThemes()) do
             print(k)
         end
     end,
     {}
-) ]]--
+)
